Test removal of blogs (5.21)
Only the creator of a blog can delete it. Test the attempt of removal by a user who didn't create the blog as well

TO CHECK: When a user likes a blog, the remove button dissappear until
refresh. (HTTP PUT request with no blog.user in the payload and the
endpoint not considering the user property may be the
keys (blog.user.name is used in the blog component to check if the user
is the owner of a blog and display the remove button. Probably is better
to check by just user.id))